third_party/blink/web_tests/external/wpt/shadow-dom/focus/focus-method-delegatesFocus-expected.txt

This is a testharness.js-based test.
PASS focus() on host with delegatesFocus, all tabindex=0
PASS focus() on host with delegatesFocus & tabindex =-1, all other tabindex=0
PASS focus() on host with delegatesFocus & no tabindex, all other tabindex=0
FAIL focus() on host with delegatesFocus & tabindex = 0, all other tabindex=-1 assert_equals: expected Element node <div tabindex="-1">aboveSlots</div> but got null
FAIL focus() on host with delegatesFocus, all without tabindex assert_equals: expected Element node <body>
<div id="host">
  <div id="slottedToSecondSlot" sl... but got Element node <div id="host">
  <div id="slottedToSecondSlot" slot="sec...
FAIL focus() on host with delegatesFocus, all tabindex=-1 assert_equals: expected Element node <div tabindex="-1">aboveSlots</div> but got null
PASS focus() on host with delegatesFocus & tabindex=0, #belowSlots with tabindex=0
FAIL focus() on host with delegatesFocus & tabindex=0, #outside with tabindex=0 assert_equals: expected Element node <body>
<div id="host" tabindex="0">
  <div id="slottedToS... but got Element node <div id="host" tabindex="0">
  <div id="slottedToSecondSl...
PASS focus() on host with delegatesFocus & tabindex=0, #aboveSlots and #belowSlots with tabindex=0
FAIL focus() on host with delegatesFocus & tabindex=0, #aboveSlots with tabindex=0 and #belowSlots with tabindex=1 assert_equals: expected Element node <div tabindex="0">aboveSlots</div> but got Element node <div tabindex="1">belowSlots</div>
PASS focus() on host with delegatesFocus & tabindex=0, #slottedToFirstSlot, #slottedToSecondSlot, #belowSlots  with tabindex=0
Harness: the test ran to completion.
